* fixed UDPInput in primitivesInspector not being a Node * clearing to 0.0 instead of -1.0 in renderer

+++ classes/tabbedListButton.cpp
@@ -79,7 +79,7 @@ void TabbedListButton::trigger(MsbObject* other){
     for (int i=0;i<(int)tabTriggerButtons.size();i++)
         tabTriggerButtons[i]->color=tabColor;
 
-    if (tabTriggerButtons.size()>currentTab)
+    if ((int)tabTriggerButtons.size()>currentTab)
         tabTriggerButtons[currentTab]->color=selectedTabColor;
 
 }
